Add files via upload
Changed the data connection to update on refresh only and then set a boolean flag to determine if a refresh had occured.  If not, then a refresh would be performed and the flag set to true to prevent further refreshes.  This make the application only refresh one time during the duration of the session.  If there are changes in inventory, lots or locations, then the application will have to be closed and then reopened.